third_party/WebKit/LayoutTests/editing/inserting/insert_br_at_end_of_block.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>x|</div>',
    'insertLineBreak',
    '<div contenteditable>x<br>|<br></div>'),
    'insert BR at end of block');

test(() => assert_selection(
    '<div contenteditable><span>x|</span></div>',
    'insertLineBreak',
    '<div contenteditable><span>x<br>|<br></span></div>'),
    'insert BR at end of block with SPAN');
</script>
